monitoring/python-dashboards/requirements.txt
Flask==3.0.0
psycopg2-binary==2.9.9
redis==5.0.1
elasticsearch==8.11.0
prometheus-client==0.19.0
pandas==2.1.4
plotly==5.17.0
dash==2.14.2
dash-bootstrap-components==1.5.0
gunicorn==21.2.0
python-dotenv==1.0.0
